+++ src/bellman_ford/bellman_ford.h
@@ -12,6 +12,6 @@
 #include <algorithm>
 #include "../fgraph/fgraph.h"
 
-t_distance_1d bellmanford(std::vector<std::vector<long int>>, int, int);
+t_distance_1d bellmanford(std::vector<std::vector<int>>, int, int);
 
 #endif //GRAFOS_BELLMAN_FORD_H
