+++ src/floyd_warshall/floyd_warshall.h
@@ -9,6 +9,6 @@
 #include <climits>
 #include "../fgraph/fgraph.h"
 
-t_distance_2d floyd_warshall(std::vector<std::vector<long int>>, int, int);
+t_distance_2d floyd_warshall(std::vector<std::vector<int>>, int, int);
 
 #endif //GRAFOS_FLOYD_WARSHALL_H
